packages/parlant-server/requirements.txt
aiofiles==24.1.0
aiopenapi3==0.8.1
aiorwlock==1.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
Authlib==1.6.4
backoff==2.2.1
bcrypt==4.3.0
boto3==1.40.37
botocore==1.40.37
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
chromadb==1.0.15
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
contextvars==2.4
contourpy==1.3.3
croniter==5.0.1
cryptography==46.0.1
cycler==0.12.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.115.12
fastmcp==2.6.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.60.0
fsspec==2025.9.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.75.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.1
humanfriendly==10.0
idna==3.10
immutables==0.21
importlib_metadata==8.4.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
jsonfinder==0.4.2
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==33.1.0
lagom==2.7.7
lazy-object-proxy==1.12.0
limits==5.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mcp==1.14.1
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.6.4
mypy==1.18.2
mypy_extensions==1.1.0
nano-vectordb==0.0.4.3
nanoid==2.0.0
networkx==3.5
numpy==2.3.3
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.109.0
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
openapi3-parser==1.1.21
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.3.2
parlant==3.0.2
parlant-client==3.0.1
pathable==0.4.4
pathspec==0.12.1
pillow==11.3.0
posthog==5.4.0
prance==25.4.8.0
propcache==0.3.2
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
s3transfer==0.14.0
scipy==1.16.2
semver==3.0.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.46.2
structlog==24.4.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.8.0
tokenizers==0.20.3
toml==0.10.2
tqdm==4.67.1
typer==0.19.2
types-aiofiles==24.1.0.20250822
types-cachetools==6.2.0.20250827
types-croniter==4.0.0.20241030
types-jsonschema==4.25.1.20250822
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
